much more logical reorganize
