+++ server/users/tests/routes/user-data.test.ts
@@ -8,25 +8,68 @@ import {
 } from "utils/test-utils";
 
 import { JERSEY_DEVIL_USER_ID } from "test/data/auth";
+import { mocked } from "jest-mock";
 import request from "supertest";
 import router from "../../routes";
+import stringify from "fast-json-stable-stringify";
 
 jest.mock("server/db-pool");
 jest.mock("server/cache");
 jest.mock("handlers/auth");
 
-describe("Tests PATCH routes of users REST API", () => {
+describe("Tests users/@me endpoint", () => {
   const server = startTestServer(router);
+
+  test("Prevents unauthorized access if no user logged in", async () => {
+    const res = await request(server.app).get(`/@me`);
+    expect(res.status).toBe(401);
+  });
+
+  test("returns data logged in user (cached)", async function () {
+    const cachedData = {
+      avatar_url: "/this_was_cached.png",
+      username: "super_cached",
+    };
+    mocked(cache().hGet).mockResolvedValueOnce(stringify(cachedData));
+    setLoggedInUser(JERSEY_DEVIL_USER_ID);
+
+    const res = await request(server.app).get(`/@me`);
+
+    expect(res.status).toBe(200);
+    expect(res.body).toEqual(cachedData);
+    expect(cache().hGet).toBeCalledTimes(1);
+    expect(cache().hGet).toBeCalledWith(CacheKeys.USER, JERSEY_DEVIL_USER_ID);
+  });
+
+  test("caches logged in user data", async function () {
+    setLoggedInUser(JERSEY_DEVIL_USER_ID);
+
+    const res = await request(server.app).get(`/@me`);
+    expect(res.status).toBe(200);
+    expect(cache().hSet).toBeCalledTimes(1);
+    expect(cache().hSet).toBeCalledWith(
+      CacheKeys.USER,
+      JERSEY_DEVIL_USER_ID,
+      // TODO: this will fail if we change what we cache. We should not
+      // rely on the whole response being cached, but declare the object ourselves.
+      stringify(res.body)
+    );
+  });
+
+  test("Returns data for the logged in user", async () => {
+    setLoggedInUser(JERSEY_DEVIL_USER_ID);
+    const res = await request(server.app).get(`/@me`);
+    expect(res.status).toBe(200);
+    expect(res.body).toEqual({
+      avatar_url: "/hannibal.png",
+      username: "jersey_devil_69",
+    });
+  });
 
   const testUserPatch = {
     username: "SnazzyNewName",
     avatarUrl:
       "https://firebasestorage.googleapis.com/v0/b/bobaboard-fb.appspot.com/o/images%2Fusers%2Favatar%2F4f5fde75-e670-4bb2-a04f-e90a5038ee34.jpeg?alt=media&token=7bdc5534-3159-4856-8d98-937cc1ee642f",
-  };
-
-  const testSettingsPatch = {
-    name: "FESTIVE_BACKGROUND_HEADER",
-    value: false,
   };
 
   test("prevents unauthorized PATCH access to the @me endpoint", async () => {
@@ -73,55 +116,6 @@ describe("Tests PATCH routes of users REST API", () => {
 
       expect(res.status).toBe(500);
       expect(res.body).toEqual({});
-    });
-  });
-
-  test("prevents unauthorized PATCH access to the @me settings endpoint", async () => {
-    await wrapWithTransaction(async () => {
-      const res = await request(server.app)
-        .patch(`/@me/settings`)
-        .send(testSettingsPatch);
-
-      expect(res.status).toBe(401);
-    });
-  });
-
-  test("updates user settings", async () => {
-    await wrapWithTransaction(async () => {
-      setLoggedInUser(JERSEY_DEVIL_USER_ID);
-      const res = await request(server.app)
-        .patch(`/@me/settings`)
-        .send(testSettingsPatch);
-
-      expect(res.status).toBe(200);
-      expect(res.body).toEqual({
-        decorations: [
-          {
-            name: "FESTIVE_BACKGROUND_HEADER",
-            type: "BOOLEAN",
-            value: false,
-          },
-        ],
-      });
-    });
-  });
-
-  test("returns an error if something went wrong while updating settings", async () => {
-    await wrapWithTransaction(async () => {
-      setLoggedInUser(JERSEY_DEVIL_USER_ID);
-      const testError = new Error("update failed");
-      jest
-        .spyOn(userQueries, "updateUserSettings")
-        .mockRejectedValueOnce(testError);
-
-      const res = await request(server.app)
-        .patch(`/@me/settings`)
-        .send(testSettingsPatch);
-
-      expect(res.status).toBe(500);
-      expect(res.body).toEqual({
-        message: `Failed to update user settings. Reason: ${testError}`,
-      });
     });
   });
 });
